[UPD] created a widget for the header to clean up code since it will be present on each screen
