+++ client/apk/lib/widgets/universal_drawer.dart
@@ -1,0 +1,182 @@
+import 'package:flutter/material.dart';
+import '../services/storage_service.dart';
+import '../screens/login_screen.dart';
+import '../screens/profile_screen.dart';
+
+class UniversalDrawer extends StatefulWidget {
+  const UniversalDrawer({Key? key}) : super(key: key);
+
+  @override
+  _UniversalDrawerState createState() => _UniversalDrawerState();
+}
+
+class _UniversalDrawerState extends State<UniversalDrawer> {
+  final TextEditingController _ipController = TextEditingController();
+
+  @override
+  void initState() {
+    super.initState();
+    _loadServerIp();
+  }
+
+  @override
+  void dispose() {
+    _ipController.dispose();
+    super.dispose();
+  }
+
+  Future<void> _loadServerIp() async {
+    final ip = await StorageService.getServerIp();
+    _ipController.text = ip;
+  }
+
+  Future<void> _saveServerIp() async {
+    await StorageService.saveServerIp(_ipController.text.trim());
+    if (mounted) {
+      ScaffoldMessenger.of(context).showSnackBar(
+        SnackBar(
+          content: Text('Server IP updated to ${_ipController.text.trim()}'),
+          backgroundColor: Colors.green,
+          duration: Duration(seconds: 2),
+        ),
+      );
+    }
+  }
+
+  Future<void> _handleLogout(BuildContext context) async {
+    await StorageService.deleteToken();
+    if (context.mounted) {
+      Navigator.pushReplacement(
+        context,
+        PageRouteBuilder(
+          pageBuilder: (context, animation, secondaryAnimation) => LoginScreen(),
+          transitionDuration: Duration.zero,
+          reverseTransitionDuration: Duration.zero,
+        ),
+      );
+    }
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Drawer(
+      backgroundColor: Colors.black,
+      child: Container(
+        decoration: BoxDecoration(
+          border: Border.all(color: Colors.white, width: 1),
+        ),
+        child: SafeArea(
+          child: Padding(
+            padding: EdgeInsets.all(20),
+            child: Column(
+              crossAxisAlignment: CrossAxisAlignment.start,
+              children: [
+                // IP Configuration Section
+                Text(
+                  'Server IP',
+                  style: TextStyle(
+                    color: Colors.white,
+                    fontSize: 18,
+                    fontWeight: FontWeight.bold,
+                  ),
+                ),
+                SizedBox(height: 10),
+                TextField(
+                  controller: _ipController,
+                  style: TextStyle(color: Colors.white),
+                  decoration: InputDecoration(
+                    hintText: 'Enter server IP',
+                    hintStyle: TextStyle(color: Colors.grey),
+                    enabledBorder: OutlineInputBorder(
+                      borderSide: BorderSide(color: Colors.white),
+                    ),
+                    focusedBorder: OutlineInputBorder(
+                      borderSide: BorderSide(color: Colors.white, width: 2),
+                    ),
+                  ),
+                ),
+                SizedBox(height: 10),
+                ElevatedButton(
+                  onPressed: _saveServerIp,
+                  style: ElevatedButton.styleFrom(
+                    backgroundColor: Colors.white,
+                    foregroundColor: Colors.black,
+                    minimumSize: Size(double.infinity, 40),
+                  ),
+                  child: Text('Save IP'),
+                ),
+                SizedBox(height: 20),
+
+                // Divider
+                Divider(color: Colors.white, thickness: 1),
+                SizedBox(height: 20),
+
+                // Services Section
+                InkWell(
+                  onTap: () {
+                    print('Services tapped');
+                    // TODO: Navigate to services screen
+                  },
+                  child: Padding(
+                    padding: EdgeInsets.all(10),
+                    child: Text(
+                      'Services',
+                      style: TextStyle(
+                        color: Colors.white,
+                        fontSize: 18,
+                        fontWeight: FontWeight.bold,
+                      ),
+                    ),
+                  ),
+                ),
+                SizedBox(height: 20),
+
+                // Divider
+                Divider(color: Colors.white, thickness: 1),
+                SizedBox(height: 20),
+
+                // Profile Section
+                InkWell(
+                  onTap: () {
+                    Navigator.push(
+                      context,
+                      MaterialPageRoute(builder: (context) => ProfileScreen()),
+                    );
+                  },
+                  child: Padding(
+                    padding: EdgeInsets.all(10),
+                    child: Text(
+                      'Profile',
+                      style: TextStyle(
+                        color: Colors.white,
+                        fontSize: 18,
+                        fontWeight: FontWeight.bold,
+                      ),
+                    ),
+                  ),
+                ),
+                SizedBox(height: 20),
+
+                // Divider
+                Divider(color: Colors.white, thickness: 1),
+
+                Spacer(),
+
+                // Logout Button at the bottom
+                ElevatedButton(
+                  onPressed: () => _handleLogout(context),
+                  style: ElevatedButton.styleFrom(
+                    backgroundColor: Colors.white,
+                    foregroundColor: Colors.black,
+                    minimumSize: Size(double.infinity, 50),
+                  ),
+                  child: Text('Logout', style: TextStyle(fontSize: 16)),
+                ),
+              ],
+            ),
+          ),
+        ),
+      ),
+    );
+  }
+}
